requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
langgraph==0.2.14
langchain-core==0.2.38
openai==1.44.1
httpx==0.27.2
mss==9.0.1
pyautogui==0.9.54
pyperclip==1.9.0
pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.1.1
PyYAML==6.0.2
